# openapi.yaml
openapi: 3.0.3
info:
  title: Portfolio Desk API
  version: 1.0.0
  description: |
    Multi-tenant office management platform API. Handles offices, leases, maintenance tickets,
    vendors, HVAC contracts, billing, and analytics.

    ## Authentication
    All protected endpoints require a JWT bearer token obtained from `/api/v1/auth/login`.
    API keys (prefixed `om_`) may be used as an alternative bearer credential.
servers:
  - url: /api/v1
    description: Primary API
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token from /auth/login or an API key prefixed with `om_`

  # -------------------------------------------------------------------------
  # Reusable schemas
  # -------------------------------------------------------------------------
  schemas:
...
    PaginationMeta:
      type: object
      properties:
        total: { type: integer }
        page: { type: integer }
        page_size: { type: integer }
        pages: { type: integer }

    # --- Error --------------------------------------------------------------
...
    Role:
      type: string
      enum: [viewer, editor, admin, ticketer, accountant]
...
    TicketStatus:
      type: string
      enum: [open, in_progress, closed]
    TicketPriority:
      type: string
      enum: [low, medium, high]
    PaymentFrequency:
      type: string
      enum: [monthly, quarterly, annually]
...
    SortOrder:
      type: string
      enum: [asc, desc]

    # --- User ---------------------------------------------------------------
    UserResponse:
      type: object
      properties:
        id: { type: string, format: uuid }
        organization_id: { type: string, format: uuid }
        email: { type: string, format: email }
        display_name: { type: string }
        auth_provider: { type: string, enum: [internal, google] }
        role: { $ref: '#/components/schemas/Role' }
        is_super_admin: { type: boolean }
        is_active: { type: boolean }
        totp_enabled: { type: boolean }
        last_login_at: { type: string, format: date-time, nullable: true }
        preferences: { type: object, nullable: true, additionalProperties: true }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
...
    RegisterRequest:
      type: object
      required: [email, display_name, password]
      properties:
        email: { type: string, format: email }
        display_name: { type: string }
        password: { type: string }
        role: { $ref: '#/components/schemas/Role' }
...
    ManagerResponse:
      type: object
      properties:
        id: { type: string, format: uuid }
        organization_id: { type: string, format: uuid }
        name: { type: string }
        email: { type: string, format: email, nullable: true }
        phone: { type: string, nullable: true }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    ManagerCreate:
      type: object
      required: [name]
      properties:
        name: { type: string }
        email: { type: string, format: email }
        phone: { type: string }
...
    OfficeResponse:
      type: object
      properties:
        id: { type: string, format: uuid }
        organization_id: { type: string, format: uuid }
        office_number: { type: integer }
        location_name: { type: string }
        region_number: { type: integer, nullable: true }
        manager_id: { type: string, format: uuid, nullable: true }
        location_type: { type: string }
        sector: { type: string }
        address_line_1: { type: string, nullable: true }
        city: { type: string }
        state: { type: string }
        zip_code: { type: string }
        phone_number: { type: string, nullable: true }
        email: { type: string, format: email, nullable: true }
        fax: { type: string, nullable: true }
        is_active: { type: boolean }
        total_sqft: { type: number, nullable: true }
        usable_sqft: { type: number, nullable: true }
        headcount_capacity: { type: integer, nullable: true }
        current_headcount: { type: integer, nullable: true }
        mail_shipping: { type: string, nullable: true }
        notes: { type: string, nullable: true }
        is_deleted: { type: boolean }
        deleted_at: { type: string, format: date-time, nullable: true }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    OfficeCreate:
      type: object
      required: [office_number, location_name, city, state, zip_code, location_type, sector]
      properties:
        office_number: { type: integer }
        location_name: { type: string }
        region_number: { type: integer }
        manager_id: { type: string, format: uuid }
        location_type: { type: string }
        sector: { type: string }
        address_line_1: { type: string }
        city: { type: string }
        state: { type: string }
        zip_code: { type: string }
        phone_number: { type: string }
        email: { type: string, format: email }
        fax: { type: string }
        is_active: { type: boolean, default: true }
        total_sqft: { type: number }
        usable_sqft: { type: number }
        headcount_capacity: { type: integer }
        current_headcount: { type: integer }
        mail_shipping: { type: string }
        notes: { type: string }
...
    PaginatedOffices:
      type: object
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/OfficeResponse' }
        meta: { $ref: '#/components/schemas/PaginationMeta' }

    # --- Lease --------------------------------------------------------------
    LeaseResponse:
      type: object
      properties:
        id: { type: string, format: uuid }
        organization_id: { type: string, format: uuid }
        office_id: { type: string, format: uuid, nullable: true }
        lease_name: { type: string }
        lessor_name: { type: string }
        lease_expiration: { type: string, format: date, nullable: true }
        lease_notice_date: { type: string, format: date, nullable: true }
        notice_given_date: { type: string, format: date, nullable: true }
        notice_period: { type: string, nullable: true }
        notice_period_days: { type: integer, nullable: true }
        payment_amount: { type: number, nullable: true }
        payment_frequency: { $ref: '#/components/schemas/PaymentFrequency' }
        annual_escalation_rate: { type: number, nullable: true }
        lease_classification: { $ref: '#/components/schemas/LeaseClassification' }
        currency: { type: string, default: USD }
        manager_id: { type: string, format: uuid, nullable: true }
        expiration_year: { type: integer }
        is_deleted: { type: boolean }
        deleted_at: { type: string, format: date-time, nullable: true }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    LeaseCreate:
      type: object
      required: [lease_name, lessor_name]
      properties:
        office_id: { type: string, format: uuid }
        lease_name: { type: string }
        lessor_name: { type: string }
        lease_expiration: { type: string, format: date }
        lease_notice_date: { type: string, format: date }
        notice_period: { type: string }
        notice_period_days: { type: integer }
        payment_amount: { type: number }
        payment_frequency: { $ref: '#/components/schemas/PaymentFrequency' }
        annual_escalation_rate: { type: number }
        lease_classification: { $ref: '#/components/schemas/LeaseClassification' }
        currency: { type: string, default: USD }
        manager_id: { type: string, format: uuid }
...
    PaginatedLeases:
      type: object
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/LeaseResponse' }
        meta: { $ref: '#/components/schemas/PaginationMeta' }
...
    MaintenanceTicketResponse:
      type: object
      properties:
        id: { type: string, format: uuid }
        organization_id: { type: string, format: uuid }
        subject: { type: string }
        description: { type: string, nullable: true }
        priority: { $ref: '#/components/schemas/TicketPriority' }
        status: { $ref: '#/components/schemas/TicketStatus' }
        category_id: { type: string, format: uuid, nullable: true }
        office_id: { type: string, format: uuid, nullable: true }
        created_by_id: { type: string, format: uuid }
        assigned_to_id: { type: string, format: uuid, nullable: true }
        closed_at: { type: string, format: date-time, nullable: true }
        is_deleted: { type: boolean }
        deleted_at: { type: string, format: date-time, nullable: true }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    MaintenanceTicketCreate:
      type: object
      required: [subject]
      properties:
        subject: { type: string }
        description: { type: string }
        priority: { $ref: '#/components/schemas/TicketPriority' }
        category_id: { type: string, format: uuid }
        office_id: { type: string, format: uuid }
        assigned_to_id: { type: string, format: uuid }
...
    PaginatedTickets:
      type: object
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/MaintenanceTicketResponse' }
        meta: { $ref: '#/components/schemas/PaginationMeta' }
...
    VendorResponse:
      type: object
      properties:
        id: { type: string, format: uuid }
        organization_id: { type: string, format: uuid }
        company_name: { type: string }
        contact_name: { type: string, nullable: true }
        contact_email: { type: string, format: email, nullable: true }
        contact_phone: { type: string, nullable: true }
        services: { type: string }
        address: { type: string, nullable: true }
        is_preferred: { type: boolean }
        office_ids:
          type: array
          items: { type: string, format: uuid }
        is_deleted: { type: boolean }
        deleted_at: { type: string, format: date-time, nullable: true }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    VendorCreate:
      type: object
      required: [company_name, services]
      properties:
        company_name: { type: string }
        contact_name: { type: string }
        contact_email: { type: string, format: email }
        contact_phone: { type: string }
        services: { type: string }
        address: { type: string }
        is_preferred: { type: boolean, default: false }
        office_ids:
          type: array
          items: { type: string, format: uuid }
...
    PaginatedVendors:
      type: object
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/VendorResponse' }
        meta: { $ref: '#/components/schemas/PaginationMeta' }

    # --- Landlord -----------------------------------------------------------
...
    LandlordCreate:
      type: object
      required: [name]
      properties:
        name: { type: string }
        contact_name: { type: string }
        contact_email: { type: string, format: email }
        contact_phone: { type: string }
        address: { type: string }
        notes: { type: string }
...
    PaginatedLandlords:
      type: object
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/LandlordResponse' }
        meta: { $ref: '#/components/schemas/PaginationMeta' }

    # --- HVAC Contract ------------------------------------------------------
...
    PaginatedHvacContracts:
      type: object
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/HvacContractResponse' }
        meta: { $ref: '#/components/schemas/PaginationMeta' }

    # --- Notification -------------------------------------------------------
...
security:
  - BearerAuth: []

# ---------------------------------------------------------------------------
# Paths
# ---------------------------------------------------------------------------
paths:
...
  /users:
    get:
      tags: [Users]
      summary: List users in organization (admin only)
      parameters:
        - { $ref: '#/components/parameters/Page' }
        - { $ref: '#/components/parameters/PageSize' }
      responses:
        '200':
          description: Paginated users
          content:
            application/json:
              schema:
                type: object
                properties:
                  items: { type: array, items: { $ref: '#/components/schemas/UserResponse' } }
                  meta: { $ref: '#/components/schemas/PaginationMeta' }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      tags: [Users]
      summary: Create user in organization (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RegisterRequest' }
      responses:
        '201':
          description: Created user
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UserResponse' }
        '402': { description: Seat limit reached }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
...
  /managers:
    get:
      tags: [Managers]
      summary: List all managers
      responses:
        '200':
          description: List of managers
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/ManagerResponse' }
    post:
      tags: [Managers]
      summary: Create manager (admin/editor)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ManagerCreate' }
      responses:
        '201':
          description: Created manager
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ManagerResponse' }
        '403': { $ref: '#/components/responses/Forbidden' }
...
  /offices:
    get:
      tags: [Offices]
      summary: List offices
      parameters:
        - { $ref: '#/components/parameters/Page' }
        - { $ref: '#/components/parameters/PageSize' }
        - { name: region_number, in: query, schema: { type: integer } }
        - { name: manager_id, in: query, schema: { type: string, format: uuid } }
        - { name: location_type, in: query, schema: { type: string } }
        - { name: sector, in: query, schema: { type: string } }
        - { name: state, in: query, schema: { type: string } }
        - { name: is_active, in: query, schema: { type: boolean } }
        - { name: search, in: query, schema: { type: string } }
        - { name: sort_by, in: query, schema: { type: string } }
        - { name: sort_order, in: query, schema: { $ref: '#/components/schemas/SortOrder' } }
      responses:
        '200':
          description: Paginated offices
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PaginatedOffices' }
    post:
      tags: [Offices]
      summary: Create office (admin/editor)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/OfficeCreate' }
      responses:
        '201':
          description: Created office
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OfficeResponse' }
        '402': { description: Office limit reached for current plan }
        '403': { $ref: '#/components/responses/Forbidden' }
...
  /leases:
    get:
      tags: [Leases]
      summary: List leases
      parameters:
        - { $ref: '#/components/parameters/Page' }
        - { $ref: '#/components/parameters/PageSize' }
        - { name: year, in: query, schema: { type: integer } }
        - { name: manager_id, in: query, schema: { type: string, format: uuid } }
        - { name: notice_status, in: query, schema: { type: string, enum: [given, not_given] } }
        - { name: expiring_within_days, in: query, schema: { type: integer } }
        - { name: sort_by, in: query, schema: { type: string } }
        - { name: sort_order, in: query, schema: { $ref: '#/components/schemas/SortOrder' } }
      responses:
        '200':
          description: Paginated leases
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PaginatedLeases' }
    post:
      tags: [Leases]
      summary: Create lease (admin/editor)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LeaseCreate' }
      responses:
        '201':
          description: Created lease
          content:
            application/json:
              schema: { $ref: '#/components/schemas/LeaseResponse' }
        '403': { $ref: '#/components/responses/Forbidden' }
...
  /maintenance-tickets:
    get:
      tags: [MaintenanceTickets]
      summary: List maintenance tickets
      parameters:
        - { $ref: '#/components/parameters/Page' }
        - { $ref: '#/components/parameters/PageSize' }
        - { name: status, in: query, schema: { $ref: '#/components/schemas/TicketStatus' } }
        - { name: priority, in: query, schema: { $ref: '#/components/schemas/TicketPriority' } }
        - { name: category_id, in: query, schema: { type: string, format: uuid } }
        - { name: office_id, in: query, schema: { type: string, format: uuid } }
        - { name: assigned_to_id, in: query, schema: { type: string, format: uuid } }
        - { name: sort_by, in: query, schema: { type: string } }
        - { name: sort_order, in: query, schema: { $ref: '#/components/schemas/SortOrder' } }
      responses:
        '200':
          description: Paginated tickets
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PaginatedTickets' }
    post:
      tags: [MaintenanceTickets]
      summary: Create ticket (admin/editor/ticketer)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/MaintenanceTicketCreate' }
      responses:
        '201':
          description: Created ticket
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MaintenanceTicketResponse' }
        '403': { $ref: '#/components/responses/Forbidden' }
...
  /vendors:
    get:
      tags: [Vendors]
      summary: List vendors
      parameters:
        - { $ref: '#/components/parameters/Page' }
        - { $ref: '#/components/parameters/PageSize' }
        - { name: search, in: query, schema: { type: string } }
        - { name: office_id, in: query, schema: { type: string, format: uuid } }
        - { name: sort_by, in: query, schema: { type: string } }
        - { name: sort_order, in: query, schema: { $ref: '#/components/schemas/SortOrder' } }
      responses:
        '200':
          description: Paginated vendors
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PaginatedVendors' }
    post:
      tags: [Vendors]
      summary: Create vendor (admin/editor)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/VendorCreate' }
      responses:
        '201':
          description: Created vendor
          content:
            application/json:
              schema: { $ref: '#/components/schemas/VendorResponse' }
        '403': { $ref: '#/components/responses/Forbidden' }
...
  /landlords:
    get:
      tags: [Landlords]
      summary: List landlords
      parameters:
        - { $ref: '#/components/parameters/Page' }
        - { $ref: '#/components/parameters/PageSize' }
        - { name: search, in: query, schema: { type: string } }
        - { name: sort_by, in: query, schema: { type: string } }
        - { name: sort_order, in: query, schema: { $ref: '#/components/schemas/SortOrder' } }
      responses:
        '200':
          description: Paginated landlords
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PaginatedLandlords' }
    post:
      tags: [Landlords]
      summary: Create landlord (admin/editor)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LandlordCreate' }
      responses:
        '201':
          description: Created landlord
          content:
            application/json:
              schema: { $ref: '#/components/schemas/LandlordResponse' }
        '403': { $ref: '#/components/responses/Forbidden' }
...
  /hvac-contracts:
    get:
      tags: [HvacContracts]
      summary: List HVAC contracts
      parameters:
        - { $ref: '#/components/parameters/Page' }
        - { $ref: '#/components/parameters/PageSize' }
        - { name: office_id, in: query, schema: { type: string, format: uuid } }
        - { name: manager_id, in: query, schema: { type: string, format: uuid } }
        - { name: status, in: query, schema: { type: string } }
      responses:
        '200':
          description: Paginated HVAC contracts
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PaginatedHvacContracts' }
    post:
      tags: [HvacContracts]
      summary: Create HVAC contract (admin/editor)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/HvacContractCreate' }
      responses:
        '201':
          description: Created contract
          content:
            application/json:
              schema: { $ref: '#/components/schemas/HvacContractResponse' }
        '403': { $ref: '#/components/responses/Forbidden' }
...
  /billing/webhooks:
    post:
      tags: [Billing]
      summary: Stripe webhook receiver (HMAC verified, public)
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object }
      responses:
        '200':
          description: Received
          content:
            application/json:
              schema:
                type: object
                properties:
                  received: { type: boolean }

  # ==========================================================================
  # API KEYS
  # ==========================================================================
  /api-keys:
    get:
      tags: [ApiKeys]
      summary: List API keys (admin only)
      responses:
        '200':
          description: API keys
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/ApiKeyResponse' }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      tags: [ApiKeys]
      summary: Create API key (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ApiKeyCreate' }
      responses:
        '201':
          description: Created key (full key value returned only once)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiKeyCreated' }
        '403': { $ref: '#/components/responses/Forbidden' }
...
  /search:
    get:
      tags: [Search]
      summary: Full-text search across offices, leases, tickets, and landlords
      parameters:
        - name: q
          in: query
          required: true
          schema: { type: string, minLength: 1, maxLength: 100 }
          description: Search query
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 50, default: 10 }
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/SearchResult' }
        '400': { $ref: '#/components/responses/BadRequest' }

  # -------------------------------------------------------------------------
  # Reusable response bodies & parameters
  # -------------------------------------------------------------------------
...